% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecific.R
\name{achillesShare}
\alias{achillesShare}
\title{Prototype inside Iris for AchillesShare}
\usage{
achillesShare(connectionDetails, cdmDatabaseSchema,
  oracleTempSchema = cdmDatabaseSchema,
  resultsDatabaseSchema = cdmDatabaseSchema, sourceName = "",
  cdmVersion = "5", vocabDatabaseSchema = cdmDatabaseSchema)
}
\description{
Prototype inside Iris for AchillesShare
}

