% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRIS.R
\name{calculate_score}
\alias{calculate_score}
\title{function used to develop ensemble score following Sahni et al. 2024 paper}
\usage{
calculate_score(test, test_name, features, sample_name = F)
}
\arguments{
\item{test}{data.frame regarding testing/validation cohort (to apply features): data.frame should have columns sample name (no format requirement), response (values binary 0 or 1), and as many CCI columns (values binary 0 or 1)}

\item{test_name}{string regarding testing/validation cohort name formatted as [cohort]_[pre or on]  (ex. Gide_on, Gide_pre, PUCH_on)}

\item{features}{LOOCohort_validation or IRIS output}

\item{sample_name}{boolean whether sample name should be included in final output or not}
}
\value{
data.frame of patient ensemble score using IRIS features
}
\description{
function used to develop ensemble score following Sahni et al. 2024 paper
}
