% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRIS.run.R
\name{createA}
\alias{createA}
\title{Construct the adjacency matrix for each tissue slice}
\usage{
createA(locationList)
}
\arguments{
\item{locationList}{a list of spatial location matrix}
}
\value{
Return a list of Ajacency matrix
}
\description{
Construct the adjacency matrix for each tissue slice
}
