% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{createIRISObject}
\alias{createIRISObject}
\title{Create the IRIS object}
\usage{
createIRISObject(
  spatial_countMat_list,
  spatial_location_list,
  sc_count,
  sc_meta,
  ct.varname,
  sample.varname,
  minCountGene = 100,
  minCountSpot = 5
)
}
\arguments{
\item{spatial_countMat_list}{A list of raw spatial resolved transcriptomics data. In each spatial transcriptomics data, it is a sparse matrix with each column is a spatial location, and each row is a gene.}

\item{spatial_location_list}{A list of spatial location data frames, with two columns representing the x and y coordinates of the spatial location (The column names must contain "x" and "y"). The rownames of this data frame should match eaxctly with the columns of the corresponding spatial_count in the spatial_countMat_list.}

\item{sc_count}{Reference scRNA-seq count data, each column is a cell and each row is a gene.}

\item{sc_meta}{A data frame of meta data providing information on each cell in the sc_count, with each row representing the cell type and/or sample information of a specific cell. The row names of this data frame should match exactly with the column names of the sc_count data}

\item{ct.varname}{character, the name of the column in metaData that provides the cell type annotation information for IRIS to perform informed domain detection.}

\item{sample.varname}{character,the name of the column in metaData that specifies the sample information. If NULL, we just use the whole as one sample.}

\item{minCountGene}{Minimum counts for each gene in the spatial count data}

\item{minCountSpot}{Minimum counts for each spatial location in the spatial count data}
}
\value{
Returns IRIS object with filtered spatial count list, location list and reference scRNA-seq dataset.
}
\description{
Create the IRIS object
}
