% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRIS.run.R
\name{createscRef}
\alias{createscRef}
\title{Construct the mean gene expression basis matrix (B), this is the faster version}
\usage{
createscRef(
  countMat,
  metadata,
  ct.select = NULL,
  ct.varname,
  sample.varname = NULL
)
}
\arguments{
\item{countMat}{single-cell RNA-seq count data as the reference}

\item{metadata}{meta data containing the cell type annotations in the scRNA-seq reference data. The rows matched with the columns of countMat}

\item{ct.select}{vector of cell type names that you are interested in to deconvolute, default as NULL. If NULL, then use all cell types provided by single cell dataset;}

\item{ct.varname}{character, the name of the column in metaData that specifies the cell type annotation information}

\item{sample.varname}{character,the name of the column in metaData that specifies the sample information. If NULL, we just use the whole as one sample.}
}
\value{
Return a list of basis (B) matrix
}
\description{
Construct the mean gene expression basis matrix (B), this is the faster version
}
