% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StudySpecific.R
\name{execute}
\alias{execute}
\title{Execute OHDSI Iris}
\usage{
execute(dbms, user = NULL, domain = NULL, password = NULL, server,
  port = NULL, cdmSchema, cdmVersion = 4, file)
}
\arguments{
\item{dbms}{The type of DBMS running on the server. Valid values are
\itemize{
  \item{"mysql" for MySQL}
  \item{"oracle" for Oracle}
  \item{"postgresql" for PostgreSQL}
  \item{"redshift" for Amazon Redshift}
  \item{"sql server" for Microsoft SQL Server}
  \item{"pdw" for Microsoft Parallel Data Warehouse (PDW)}
  \item{"netezza" for IBM Netezza}
}}

\item{user}{The user name used to access the server. If the user is not specified for SQL Server,
Windows Integrated Security will be used, which requires the SQL Server JDBC drivers
to be installed.}

\item{domain}{(optional) The Windows domain for SQL Server only.}

\item{password}{The password for that user}

\item{server}{The name of the server}

\item{port}{(optional) The port on the server to connect to}

\item{cdmSchema}{Schema name where your patient-level data in OMOP CDM format resides}

\item{cdmVersion}{Define the OMOP CDM version used:  currently support 4 and 5.  Default = 4}

\item{file}{(Optional) Name of local file to place results; makre sure to use forward slashes (/)}
}
\value{
Study results are placed in CSV format files in specified local folder and returned
as an R object class \code{OhdsiStudy} when sufficiently small.  The properties of an
\code{OhdsiStudy} may differ from study to study.
}
\description{
Execute OHDSI Iris
}
\details{
This function executes OHDSI Iris and emails the results directly to Vojtech Huser.
Iris ...  (GIVE SOME DETAILS).
}
\examples{
\dontrun{
# Run study
execute(dbms = "postgresql",
        user = "joebruin",
        password = "supersecret",
        server = "myserver",
        cdmSchema = "cdm_schema",
        cdmVersion = 4)

# Email result file
email(from = "collaborator@ohdsi.org",
      dataDescription = "CDM4 Simulated Data")
}
}

