% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecific.R
\name{executePart}
\alias{executePart}
\title{Execute Iris Part  (testing new components)}
\usage{
executePart(part = 1, connectionDetails, cdmVersion = 5,
  file = "iris2results.ohdsi")
}
\details{
This function executes a logical part of Iris (easier for testing of multiple parts)
}

