% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRIS.run.R
\name{getConcated}
\alias{getConcated}
\title{Construct the concatenated matrix from V across tissue_slices}
\usage{
getConcated(VList, locationList, sliceNames)
}
\arguments{
\item{VList}{a list of V matrix across all tissue slices}

\item{locationList}{a list of location matrix matched with the spatial locations in the VList}

\item{sliceNames}{a vector of slice names corresponding to the input tissue slices}
}
\value{
Return a list of Ajacency matrix
}
\description{
Construct the concatenated matrix from V across tissue_slices
}
