% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRIS.run.R
\name{getMu_and_STR}
\alias{getMu_and_STR}
\title{get the Mu and spatial domain label list}
\usage{
getMu_and_STR(
  CommonCellTypes,
  TotalClusters,
  MatrixCombined,
  nSlices,
  sliceNames
)
}
\arguments{
\item{CommonCellTypes}{cell types in the reference matrix}

\item{TotalClusters}{number of total spatial odmains across all tissue slices}

\item{MatrixCombined}{a concatenated matrix of V matrix across tissue slices}

\item{nSlices}{number of total slices analyzed}

\item{sliceNames}{names of the tissue slices in the count list}
}
\value{
Return a list of Mu and the spatial domain labels for each tissue slice
}
\description{
get the Mu and spatial domain label list
}
