% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{class}
\name{IRIS-class}
\alias{IRIS-class}
\title{Each IRIS object has a number of slots which store information. Key slots to access
are listed below.}
\description{
Each IRIS object has a number of slots which store information. Key slots to access
are listed below.
}
\section{Slots}{

\describe{
\item{\code{countList}}{A list of spatial count data matrix. Each spatial count data with rows as the genes and column as measured spatial locations}

\item{\code{locationList}}{A list of spatial location matrix. Each spatial location with rows as the measured spatial locations matched with those in the spatial_countMat_list and columns as theccccczzzz}

\item{\code{spatialDomain}}{A meta data frame of spatial domain labels detected by IRIS.}

\item{\code{project}}{The name of the project, default is "IRIS".}

\item{\code{IRIS_Prop}}{A meta data frame of concatenated V matrix estimated by IRIS.}

\item{\code{internal_info}}{The neccessary information that are used by IRIS computation.}
}}

