% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRIS.R
\name{IRIS}
\alias{IRIS}
\title{Immunotherapy Resistant cell-cell Interaction Scanner}
\usage{
IRIS(s2_cohort, s2_name, s1_cohort, direction)
}
\arguments{
\item{s2_cohort}{s2_cohort: data.frame of pre-treatment cohort to be used for step2 feature selection; data.frame should have columns sample name (no format requirement), response (values binary 0 or 1), and as many CCI columns (values binary 0 or 1)}

\item{s2_name}{string of cohort name used for step 2 feature selection. names formatted as [cohort]_[pre]. Input random [cohort] name if you want to reutilize step 1 pre-treatment cohort in step 2 (i.e. RANDOM_pre)}

\item{s1_cohort}{named list of (data.frame) pre/on-treatment cohort(s) to be used for step1 feature selection; names formatted as [cohort]_[pre/on] (requires at least one pre and one on-rx cohort); data.frame should have columns sample name (no format requirement), response (values binary 0 or 1), and as many CCI columns (values binary 0 or 1)}

\item{direction}{string ('RDI' or 'RUI') corresponding to searching for RDIs or RUIs}
}
\value{
vector of selected interactions (Lcell_Rcell_Ligand_Receptor) from IRIS
}
\description{
function to infer upregulated or downregulated cell-cell interactions that mediate resistance
}
