% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{IRIS.visualize.eachProp}
\alias{IRIS.visualize.eachProp}
\title{Visualize the spatial distribution of cell type proportion}
\usage{
IRIS.visualize.eachProp(
  proportion,
  spatial_location,
  ct.visualize = ct.visualize,
  colors = c("lightblue", "lightyellow", "red"),
  numCols
)
}
\arguments{
\item{proportion}{Data frame, cell type proportion estimated by IRIS in either original resolution or enhanced resolution.}

\item{spatial_location}{Data frame, spatial location information.}

\item{ct.visualize}{Vector of selected cell type names that are interested to visualize}

\item{colors}{Vector of color names that you want to use, if NULL, we will use the default color scale c("lightblue","lightyellow","red")}

\item{numCols}{Numeric, number of columns in the figure panel, it depends on the number of cell types you want to visualize.}
}
\value{
Returns a ggplot2 figure.
}
\description{
Visualize the spatial distribution of cell type proportion
}
