% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IRIS_Marker_iter}
\alias{IRIS_Marker_iter}
\title{IRIS_ref function}
\usage{
IRIS_Marker_iter(XList, BIn, AList, VList, StrList, MuIn)
}
\arguments{
\item{XList}{The input list of normalized spatial data}

\item{BIn}{The input list of cell type specific basis matrix B}

\item{AList}{The constructed Ajacency matrix for each tissue slice}

\item{VList}{List of initial matrix of cell type compositions V}

\item{StrList}{List of initial spatial domains for each tissue slice}

\item{MuIn}{Initial matrix of mean cell type proportion across slices}
}
\value{
A list
}
\description{
IRIS_ref function
}
