% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRIS.run.R
\name{IRIS_spatial}
\alias{IRIS_spatial}
\title{Accurate and Efficient Integrative Reference-Informed Spatial Domain Detection for Spatial Transcriptomics}
\usage{
IRIS_spatial(IRIS_object, numCluster)
}
\arguments{
\item{IRIS_object}{IRIS object create by the createIRISObject function}

\item{numCluster}{A numeric value indicating the number of clusters}
}
\value{
Returns a IRIS object with domain labels stored in IRIS_object@spatial_domain for each tissue slice.
}
\description{
Accurate and Efficient Integrative Reference-Informed Spatial Domain Detection for Spatial Transcriptomics
}
