% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRIS.run.R
\name{kmeansFunc_Initialize}
\alias{kmeansFunc_Initialize}
\title{Initialize the spatial domain labels}
\usage{
kmeansFunc_Initialize(data, k)
}
\arguments{
\item{data}{Contatnated matrix from all tissue slices.}

\item{k}{number of spatial domains}
}
\value{
a list of the kmeans clustering results
}
\description{
Initialize the spatial domain labels
}
