% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRIS.run.R
\name{kmeansFunc_Iter}
\alias{kmeansFunc_Iter}
\title{Cluster the spatial domain labels in each iterartion}
\usage{
kmeansFunc_Iter(data, k, centers_old = NULL)
}
\arguments{
\item{data}{Contatnated matrix from all tissue slices.}

\item{k}{number of spatial domains}

\item{centers_old}{centers from last iteration}
}
\value{
a list of the kmeans clustering results and the centers for this iteration
}
\description{
Cluster the spatial domain labels in each iterartion
}
