% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRIS.R
\name{LOOC_validation}
\alias{LOOC_validation}
\title{Leave One Out Cohort Validation}
\usage{
LOOC_validation(test_name, s1_cohort, s2_cohort, direction)
}
\arguments{
\item{test_name}{string regarding testing/validation cohort name formatted as [cohort]_[pre or on]  (ex. Gide_on, Gide_pre, PUCH_on)}

\item{s1_cohort}{named list of (data.frame) pre/on-treatment cohort(s) to be used for step1 feature selection; names formatted as [cohort]_[pre/on] (requires at least one pre and one on-rx cohort);  data.frame should have columns sample name (no format requirement), response (values binary 0 or 1), and as many CCI columns (values binary 0 or 1)}

\item{s2_cohort}{s2_cohort: named list of (data.frame) pre-treatment cohort(s) to be used for step2 feature selection; names formatted as [cohort]_[pre]; data.frame should have columns sample name (no format requirement), response (values binary 0 or 1), and as many CCI columns (values binary 0 or 1)}

\item{direction}{string ('RDI' or 'RUI') corresponding to searching for RDIs or RUIs}
}
\value{
named list of string of RDIs or RUIs interactions; name correspond to cohort used in step 2 feature selection
}
\description{
function used to perform Leave One Out Cohort (LOOC) validation for IRIS
}
