% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primr_get.R
\name{primr_get}
\alias{primr_get}
\title{Get refuge survey data from PRIMR using web services}
\usage{
primr_get(
  max = 1000,
  response_order = "id",
  cost_code = NULL,
  type = NULL,
  status = NULL,
  species = NULL,
  output = "list",
  ...
)
}
\arguments{
\item{max}{The number of rows to return. Default is \code{1000}.}

\item{response_order}{The order of the returned data frame By default, response_order=id.}

\item{cost_code}{The unique refuge cost code. Use \%25 as a wild card (e.g., FF07\%25 for all Region 7 surveys)}

\item{type}{An exact matching (case insensitive) survey type name. Accepted values include:
\itemize{
 \item{Inventory}{}
 \item{Coop Inventory}{}
 \item{Baseline Monitoring}{}
 \item{Coop Baseline Monitoring}{}
 \item{Monitoring to Inform Management}{}
 \item{Coop Monitoring to Inform Management}{}
 \item{Research}{}
 \item{Coop Research}{}
 }}

\item{status}{An exact matching survey status name. Accepted values include:
\itemize{
 \item{Current}{}
 \item{Expected}
 \item{Historic}
 \item{Future}
}}

\item{species}{the species name, matched against any listed or non-listed species scientific name or common name, case insensitively}

\item{output}{The format of the data to be returned. Acceptable values include:
\itemize{
 \item{df}{Results will be returned as a flattened data frame}
 \item{list}{Results will be returned as a list of data frames for each refuge}
 }}

\item{...}{arguments to pass}
}
\value{
a data frame containing FWS refuge survey metadata
}
\description{
Gets data from the FWS Refuge PRIMR database using a web service API (https://my.usgs.gov/confluence/display/FEAH/Survey+Web+Service+API)
}
\examples{
\dontrun{
primr_get(cost_code = "FF01RTBL00")
}
}
\author{
McCrea Cobb \email{mccrea_cobb@fws.gov}
}
