% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primr_get.R
\name{primr_get_ak}
\alias{primr_get_ak}
\title{Get PRIMR data for Alaska Refuges using web services}
\usage{
primr_get_ak(status = NULL, saveit = FALSE, filedir, ...)
}
\arguments{
\item{status}{the status of the surveys to return. Accepted values include:
\itemize{
 \item{Current (the default value)}
 \item{Expected}
 \item{Historic}
 \item{Future}
}}

\item{saveit}{a logical value indicating whether the returned data frame be saved}

\item{filedir}{a connection or the name of the file where the data will be saved}

\item{...}{inherited arguments}
}
\value{
a data frame containing PRIMR survey data for the Alaska region
}
\description{
Gets survey data for Alaska Refuges from the FWS Refuge PRIMR database using an API
}
\examples{
\dontrun{
get_primr_ak(status = "Current", saveit = TRUE)
}
}
\author{
McCrea Cobb \email{mccrea_cobb@fws.gov}
}
