% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_iris.R
\name{query_iris}
\alias{query_iris}
\title{Connect to and query the FWS IRIS Warehouse}
\usage{
query_iris(schema, tbl_name, ...)
}
\arguments{
\item{schema}{The schema of the database in the IRIS Warehouse}

\item{tbl_name}{The name of the desired table in the IRIS Warehouse}

\item{...}{inherited arguments}
}
\value{
A SQL query list returned from \code{dplyr::tbl}
}
\description{
Returns a query of the IRIS Data Warehouse with columns \code{column_size==0} ordered last. \strong{Requires a VPN connection.}
This allow the table to be read into R using ODBC/DBI and addresses the "Invalid descriptor error" (https://github.com/r-dbi/odbc/issues/309).
}
\examples{
\dontrun{
query_iris("Refuges", "DimSurvey")
}
}
\author{
McCrea Cobb \email{mccrea_cobb@fws.gov}
}
