% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperSqlFunctions.R
\name{renderTranslateSql}
\alias{renderTranslateSql}
\title{render parametized SQL using parameters from the dataSetConnectionDetails object}
\usage{
renderTranslateSql(dataSetConnectionDetails, parameterizedSql, ...,
  oracleTempSchema = NULL)
}
\description{
render parametized SQL using parameters from the dataSetConnectionDetails object
}

