% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecific.R
\name{runParam}
\alias{runParam}
\title{Query parametized SQL using dataSetConnectionDetails object}
\usage{
runParam(dataSetConnectionDetails, parameterizedSql, ...,
  oracleTempSchema = NULL)
}
\description{
Query parametized SQL using dataSetConnectionDetails object
}

