% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sc_QC}
\alias{sc_QC}
\title{Quality control of scRNA-seq count data}
\usage{
sc_QC(
  sc_count,
  sc_meta,
  ct.varname,
  sample.varname = NULL,
  min.cells = 0,
  min.genes = 0
)
}
\arguments{
\item{sc_count}{Raw scRNAseq count data, each column is a cell and each row is a gene.}

\item{sc_meta}{data frame, metaData with "ct.varname" specify the cell type annotation information and "sample.varname" specify the sample information}

\item{ct.varname}{character, the name of the column in metaData that specifies the cell type annotation information}

\item{sample.varname}{character,the name of the column in metaData that specifies the sample information. If NULL, we just use the whole as one sample.}

\item{min.cells}{numeric, we filtered out the non-expressed cells.}

\item{min.genes}{numeric we filtered out the non-expressed genes}
}
\value{
Return the filtered scRNA-seq data and meta data stored in a S4 class (SingleCellExperiment)
}
\description{
Quality control of scRNA-seq count data
}
