% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecific.R, R/helperSqlFunctions.R
\name{createDataSetConnectionDetails}
\alias{createDataSetConnectionDetails}
\title{DataSetConnection object to facilitate passing additional server parameters to parametized SQL}
\usage{
createDataSetConnectionDetails(connectionDetails, cdmDatabaseSchema,
  resultsDatabaseSchema, vocabDatabaseSchema = cdmDatabaseSchema,
  cdmVersion = "5", oracleTempSchema = NULL)

createDataSetConnectionDetails(connectionDetails, cdmDatabaseSchema,
  resultsDatabaseSchema, vocabDatabaseSchema = cdmDatabaseSchema,
  cdmVersion = "5", oracleTempSchema = NULL)
}
\description{
DataSetConnection object to facilitate passing additional server parameters to parametized SQL

DataSetConnection object to facilitate passing additional server parameters to parametized SQL
}

